requirements.txt
altgraph==0.15
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
flake8==3.5.0
future==0.16.0
Gooey==0.9.5.2
jdcal==1.3
macholib==1.9
matplotlib==2.1.1
mccabe==0.6.1
numpy==1.14.0
olefile==0.45.1
openpyxl==2.4.9
pdfminer.six==20170720
pefile==2017.11.5
Pillow==4.3.0
psutil==5.4.2
pycodestyle==2.3.1
pycryptodome==3.6.1
pyflakes==1.6.0
PyInstaller==3.3.1
pyparsing==2.2.0
PyPDF2==1.26.0
pypiwin32==219
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
tqdm==4.19.5
wxPython==4.0.0b1
